feat: retire second-wave structural debt in cli, cache, and startup surfaces
