+++ src/vllm-sr/cli/templates/grafana_panel_factories.py
@@ -1,0 +1,173 @@
+"""Grafana panel JSON factories for Prometheus-backed dashboards."""
+
+
+def create_stat_panel(title, expr, unit="short", x=0, y=0, w=6, h=6, panel_id=1):
+    """Create a stat panel"""
+    return {
+        "datasource": {"type": "prometheus", "uid": "${DS_PROMETHEUS}"},
+        "fieldConfig": {
+            "defaults": {
+                "color": {"mode": "thresholds"},
+                "mappings": [],
+                "thresholds": {
+                    "mode": "absolute",
+                    "steps": [{"color": "green", "value": None}],
+                },
+                "unit": unit,
+            }
+        },
+        "gridPos": {"h": h, "w": w, "x": x, "y": y},
+        "id": panel_id,
+        "options": {
+            "colorMode": "value",
+            "graphMode": "area",
+            "justifyMode": "auto",
+            "orientation": "auto",
+            "reduceOptions": {"calcs": ["lastNotNull"], "fields": "", "values": False},
+            "textMode": "auto",
+        },
+        "pluginVersion": "11.5.1",
+        "targets": [
+            {
+                "datasource": {"type": "prometheus", "uid": "${DS_PROMETHEUS}"},
+                "expr": expr,
+                "refId": "A",
+            }
+        ],
+        "title": title,
+        "type": "stat",
+    }
+
+
+def create_timeseries_panel(
+    title, targets, x=0, y=0, w=12, h=8, panel_id=1, unit="short"
+):
+    """Create a time series panel"""
+    return {
+        "datasource": {"type": "prometheus", "uid": "${DS_PROMETHEUS}"},
+        "fieldConfig": {
+            "defaults": {
+                "color": {"mode": "palette-classic"},
+                "custom": {
+                    "axisBorderShow": False,
+                    "axisCenteredZero": False,
+                    "axisColorMode": "text",
+                    "axisLabel": "",
+                    "axisPlacement": "auto",
+                    "barAlignment": 0,
+                    "drawStyle": "line",
+                    "fillOpacity": 10,
+                    "gradientMode": "none",
+                    "hideFrom": {"tooltip": False, "viz": False, "legend": False},
+                    "insertNulls": False,
+                    "lineInterpolation": "linear",
+                    "lineWidth": 1,
+                    "pointSize": 5,
+                    "scaleDistribution": {"type": "linear"},
+                    "showPoints": "never",
+                    "spanNulls": False,
+                    "stacking": {"group": "A", "mode": "none"},
+                    "thresholdsStyle": {"mode": "off"},
+                },
+                "mappings": [],
+                "thresholds": {
+                    "mode": "absolute",
+                    "steps": [{"color": "green", "value": None}],
+                },
+                "unit": unit,
+            }
+        },
+        "gridPos": {"h": h, "w": w, "x": x, "y": y},
+        "id": panel_id,
+        "options": {
+            "legend": {
+                "calcs": [],
+                "displayMode": "list",
+                "placement": "bottom",
+                "showLegend": True,
+            },
+            "tooltip": {"mode": "multi", "sort": "none"},
+        },
+        "pluginVersion": "11.5.1",
+        "targets": targets,
+        "title": title,
+        "type": "timeseries",
+    }
+
+
+def create_row_panel(title, y=0, panel_id=100):
+    """Create a row panel"""
+    return {
+        "collapsed": False,
+        "gridPos": {"h": 1, "w": 24, "x": 0, "y": y},
+        "id": panel_id,
+        "panels": [],
+        "title": title,
+        "type": "row",
+    }
+
+
+def create_target(expr, legend="", ref_id="A"):
+    """Create a query target"""
+    return {
+        "datasource": {"type": "prometheus", "uid": "${DS_PROMETHEUS}"},
+        "expr": expr,
+        "legendFormat": legend,
+        "refId": ref_id,
+    }
+
+
+def create_bar_chart_panel(
+    title, targets, x=0, y=0, w=24, h=8, panel_id=1, unit="short"
+):
+    """Create a bar chart panel"""
+    return {
+        "datasource": {"type": "prometheus", "uid": "${DS_PROMETHEUS}"},
+        "fieldConfig": {
+            "defaults": {
+                "color": {"mode": "palette-classic"},
+                "custom": {
+                    "axisBorderShow": False,
+                    "axisCenteredZero": False,
+                    "axisColorMode": "text",
+                    "axisLabel": "",
+                    "axisPlacement": "auto",
+                    "barAlignment": 0,
+                    "drawStyle": "bars",
+                    "fillOpacity": 80,
+                    "gradientMode": "none",
+                    "hideFrom": {"tooltip": False, "viz": False, "legend": False},
+                    "insertNulls": False,
+                    "lineInterpolation": "linear",
+                    "lineWidth": 1,
+                    "pointSize": 5,
+                    "scaleDistribution": {"type": "linear"},
+                    "showPoints": "never",
+                    "spanNulls": False,
+                    "stacking": {"group": "A", "mode": "none"},
+                    "thresholdsStyle": {"mode": "off"},
+                },
+                "mappings": [],
+                "thresholds": {
+                    "mode": "absolute",
+                    "steps": [{"color": "green", "value": None}],
+                },
+                "unit": unit,
+            }
+        },
+        "gridPos": {"h": h, "w": w, "x": x, "y": y},
+        "id": panel_id,
+        "options": {
+            "legend": {
+                "calcs": ["lastNotNull"],
+                "displayMode": "table",
+                "placement": "right",
+                "showLegend": True,
+            },
+            "tooltip": {"mode": "multi", "sort": "desc"},
+        },
+        "pluginVersion": "11.5.1",
+        "targets": targets,
+        "title": title,
+        "type": "timeseries",
+    }
